pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pathfinder"
version = "0.2.0"
description = "Pathfinder — AI-driven architecture management using Structurizr and C4"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
]

[project.scripts]
pathfinder = "pathfinder.cli.main:cli"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[tool.hatch.build.targets.wheel]
packages = ["pathfinder"]

[tool.hatch.build.targets.wheel.force-include]
"pathfinder/skills" = "pathfinder/skills"
"pathfinder/agents" = "pathfinder/agents"

[tool.pytest.ini_options]
testpaths = ["tests"]
